requirements.txt
astroid==2.2.5
boto3==1.9.227
botocore==1.12.227
dj-database-url==0.5.0
Django==2.2.3
django-cors-headers==3.0.2
django-filter==2.1.0
django-haystack==2.8.1
django-heroku==0.3.1
django-randomslugfield==0.3.0
django-rest-auth==0.9.5
django-s3-upload==0.2.1
djangorestframework==3.9.4
docutils==0.15.2
fake-factory==9999.9.9
Faker==2.0.0
gunicorn==20.0.2
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.4
wrapt==1.11.2
